be/requirements.txt
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
redis==4.5.5
